Added animal images fixed CSS reference, updated IndexController
Added an images folder to static context of the project, moved css to a static context in the project, re-introduced css links to the view files and removed local css. Updated visitor view to display animal image previews. Added a function to index controller to retrieve a link for animals in the visitor view. Fixed visitor view to display cost of entry and changed styling.